move implementation-specific log writer methods out of base
Moves the implementation-specific log writer methods out of the base LogWriter class. This allows descendants of LogWriter to use the LogEvent object which includes more details than were previously passed to the abstract `_write` method.

While the `_getLogPrefix` member of LogWriter could previously access the LogEvent object, not all log writers are going to be building a string, so the idea of a prefix does not always apply.